docs: add tier specific premium-submitted pages with fixed lead values
The premium form no longer stores its values in a local RxDB document,
so premium-submitted.tsx imported FORM_VALUE_DOCUMENT_ID and
FormValueDocData that no longer exist and the premium_lead event
never fired. Replace the local-document lookup with tier specific
thank-you pages (premium-submitted-pro, premium-submitted-pro-plus)
that report the lead value from the fixed tier pricing.
